svm: conformance: repackage instr harness
