+++ svm/src/conformance/instr/context.rs
@@ -1,20 +1,15 @@
-//! Instruction context and effects (input + output).
+//! Instruction context (input).
 
 #[cfg(feature = "conformance")]
 use {
-    super::{
-        account_state::{account_from_proto, account_to_proto},
-        feature_set::feature_set_from_proto,
-    },
-    protosol::protos::{InstrContext as ProtoInstrContext, InstrEffects as ProtoInstrEffects},
+    crate::conformance::{account_state::account_from_proto, feature_set::feature_set_from_proto},
+    protosol::protos::InstrContext as ProtoInstrContext,
     solana_instruction::AccountMeta,
 };
 use {
-    solana_account::Account,
-    solana_instruction::{Instruction, error::InstructionError},
+    solana_account::Account, solana_instruction::Instruction,
     solana_program_runtime::execution_budget::DEFAULT_INSTRUCTION_COMPUTE_UNIT_LIMIT,
-    solana_pubkey::Pubkey,
-    solana_svm_feature_set::SVMFeatureSet,
+    solana_pubkey::Pubkey, solana_svm_feature_set::SVMFeatureSet,
 };
 
 /// Inputs to a single instruction.
@@ -91,58 +86,6 @@ impl From<ProtoInstrContext> for InstrContext {
             accounts,
             instruction,
             cu_avail,
-        }
-    }
-}
-
-/// Represents the effects of a single instruction.
-pub struct InstrEffects {
-    pub result: Option<InstructionError>,
-    pub custom_err: Option<u32>,
-    pub resulting_accounts: Vec<(Pubkey, Account)>,
-    pub cu_avail: u64,
-    pub return_data: Vec<u8>,
-    pub logs: Vec<String>,
-}
-
-impl InstrEffects {
-    /// Returns the resulting account for the given pubkey, if it exists.
-    pub fn get_account(&self, pubkey: &Pubkey) -> Option<&Account> {
-        self.resulting_accounts
-            .iter()
-            .find(|(pk, _)| pk == pubkey)
-            .map(|(_, acc)| acc)
-    }
-}
-
-#[cfg(feature = "conformance")]
-impl From<InstrEffects> for ProtoInstrEffects {
-    fn from(value: InstrEffects) -> Self {
-        let InstrEffects {
-            result,
-            custom_err,
-            resulting_accounts,
-            cu_avail,
-            return_data,
-            ..
-        } = value;
-
-        Self {
-            result: result
-                .as_ref()
-                .map(|error| {
-                    let serialized_err = bincode::serialize(error).unwrap();
-                    i32::from_le_bytes((&serialized_err[0..4]).try_into().unwrap())
-                        .saturating_add(1)
-                })
-                .unwrap_or_default(),
-            custom_err: custom_err.unwrap_or_default(),
-            modified_accounts: resulting_accounts
-                .into_iter()
-                .map(account_to_proto)
-                .collect(),
-            cu_avail,
-            return_data,
         }
     }
 }
