+++ svm/src/conformance/mod.rs
@@ -2,8 +2,7 @@
 
 #[cfg(feature = "conformance")]
 pub mod account_state;
-pub mod context;
 #[cfg(feature = "conformance")]
 pub mod feature_set;
-pub mod harness;
+pub mod instr;
 pub mod programs;
